Fix crash caused by ScrollState handling in dropdown menu
