+++ table-format/src/commonMain/kotlin/ua/wwind/table/format/data/EditFormatRule.kt
@@ -1,0 +1,10 @@
+package ua.wwind.table.format.data
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+internal data class EditFormatRule<E : Enum<E>, FILTER>(
+    val index: Int,
+    val item: TableFormatRule<E, FILTER>,
+    val isNew: Boolean = false,
+)
